It's building now.  Next: Test particle sorting
No particle movement or collision yet.  I want to get the sorting to
work first with something simple like creating 1,000,000 particles (same
as the OriginalData structures in the preceding project GpuRadixSort on
which this project is based), setting up an incrementing value in each
particle (pick a member, set particle 0's to 0, particle 1's to 1, ...,
particle 537th's to 537, etc.), scramble the particles, then sort them
again.  After that I shall move on to sorting via Morton Codes and
verify that.