I fixed it
I learned (was reminded?) that a 3D vector's W value is used as a
scaling factor.  When I normalized the whole vec4 particle position and
W was 0 (because ParticleUpdate.comp isn't running yet to give the
particles' positions a W = 1), then gl_Position got a vec4 with W = 0.
It looks like the position scales with the inverse of W, so as W goes to
0, the scaling goes to infinity.