Fixed velocity "random" issue
It seems that it wasn't good to use the blend alpha value for variable
position for the blend alpha value for variable velocity.  Now velocity
gets its own blend alpha in both bar and point emitters (I'm not using
the latter right now, but I made the fix there anyway).