FIXED IT!
I wasn't synchronizing correctly.  I was supposed to
(1) glFenceSync(...) to create a sync object and inject it into the
pipline
(2) glClientWaitSync(...) until the condition was satisfied
(3) write to buffer

Ditto for the read

I was calling glFenceSync(...), reading from/writing to the buffer, THEN
calling glClientWaitSync(...).  Whoops.  Lessons were learned, and it
works!