Yup, my Morton Code was off.
Normalizing was meant to reduce the particle position to the range
[-1,+1], but I overlooked the fact that it sets the particle vector to
be a length of exactly 1.  I figured this out by working on an earlier
demo and saw that my Z order curve was messed up.

I still need to figure out a way to get the particle range into
PositionToMortonCode(...) or have a way to get it to the function that
calls it.  Either PositionToMortonCode(...) or the function that calls
it is going to have to divide the particle position by the magnitude of
the total range.  Normalize is no good.