Created a ParticleRegionBoundaries compute "header".
I was incorrectly preparing the argument vector of
PositionToMortonCode(...) for calculation.  I was normalizing it instead
of dividing it by the valid range.  I needed to get the particle range
into PositionToMortonCode.comp, so I chose to use another compute
"header".  I could have done uniforms, and I may change back to that in
the future.