Trying to check if the frame rate drop was due to the now double size of the ParticleBuffer.
Nope.  Increasing particle count from 50,000 to 100,000 dropped the
frame rate nearly in half, so I was curious if the halving of the frame
rate by doubling of the ParticleBuffer's capacity all that was
happening.  Nope.  I stopping the Draw() method from doing anything and
the frame rate was still crap.