Nope, this didn't help.
I was hoping that making a Particle structure that was composed of
primitives only, and therefore avoided the 16-byte forced alignment that
is created when using native GLSL structures like vec4, would allow
Particle structures to be shuffled arond rapidly in memory in each loop
in which SortParticlesByPrefixSum.comp is invoked.  But nope.  Still
have a framerate down in the mid-low 20s.