Re-enabled collision detection and coloring, and dropped the max particle count to 20k.
This looks like it is working, but the collisions are not happening like
they should.  Bounding volumes are required.  See next project,
"2D_GPU_PCollisionWithBvh".

Three optimization attempts were made on branches:
- GodShaderSortAttempt: Hoping that binding a single shader and invoking
it with a "stage" enum would improve performance.  It dropped
performance.
- TrySortingWithoutIntermediateData: Hoping that avoiding the middle-man
would make things easier to maintain.  It would, but at the cost of a
nearly 2/3 performance drop.
- TryReplacingParticleVec4sWithPrimitives: I was thinking that the
performance drop of the "without intermediate data" branch was due to
memory bank conflicts while writing the GLSL native structure vec4,
which is 16byte.  I replaced the vec4s with primitives, but no
performance gain.